+++ src/headers/src/ProjectInfoPopup.cpp
@@ -1,3 +1,5 @@
+#include "../../Debugger.hpp"
+
 #include "../ProjectInfoPopup.hpp"
 
 #include "../../incl/Avalanche.hpp"
@@ -5,6 +7,7 @@
 #include <Geode/Geode.hpp>
 
 #include <Geode/ui/General.hpp>
+#include <Geode/ui/GeodeUI.hpp>
 #include <Geode/ui/Notification.hpp>
 
 #include <Geode/utils/web.hpp>
@@ -13,8 +16,8 @@
 #include <Geode/binding/CCMenuItemSpriteExtra.hpp>
 #include <Geode/binding/ButtonSprite.hpp>
 
-using namespace avalanche;
 using namespace geode::prelude;
+using namespace avalanche;
 
 ProjectInfoPopup* ProjectInfoPopup::create() {
   auto ret = new ProjectInfoPopup;
@@ -63,10 +66,10 @@ void ProjectInfoPopup::infoPopup(CCObject*) {
     "OK", "Watch",
     [this](auto, bool btn2) {
       if (btn2) {
-        log::info("Opening showcase link in browser: {}", this->m_avalProject.showcase_url);
+        AVAL_LOG_INFO("Opening showcase link in browser: {}", this->m_avalProject.showcase_url);
         web::openLinkInBrowser(this->m_avalProject.showcase_url);
       } else {
-        log::debug("User clicked OK");
+        AVAL_LOG_DEBUG("User clicked OK");
       }; },
     true);
 };
@@ -83,10 +86,10 @@ void ProjectInfoPopup::onFameInfo(CCObject*) {
     "OK", "Learn More",
     [](auto, bool btn2) {
       if (btn2) {
-        log::info("Opening Hall of Fame link in browser");
+        AVAL_LOG_INFO("Opening Hall of Fame link in browser");
         web::openLinkInBrowser(URL_AVALANCHE);
       } else {
-        log::debug("User clicked OK");
+        AVAL_LOG_DEBUG("User clicked OK");
       }; },
     true);
 };
@@ -103,12 +106,17 @@ void ProjectInfoPopup::onPlayShowcase(CCObject*) {
     "Cancel", "Watch",
     [this](auto, bool btn2) {
       if (btn2) {
-        log::info("Opening showcase link in browser: {}", this->m_avalProject.showcase_url);
+        AVAL_LOG_INFO("Opening showcase link in browser: {}", this->m_avalProject.showcase_url);
         web::openLinkInBrowser(this->m_avalProject.showcase_url);
       } else {
-        log::debug("User clicked Cancel");
+        AVAL_LOG_DEBUG("User clicked Cancel");
       }; },
     true);
+};
+
+void ProjectInfoPopup::settingsPopup(CCObject*) {
+  AVAL_LOG_INFO("Opening settings popup");
+  openSettingsPopup(getMod());
 };
 
 bool ProjectInfoPopup::setup() {
@@ -123,7 +131,7 @@ bool ProjectInfoPopup::setup() {
 
   // for buttons to work
   m_overlayMenu = CCMenu::create();
-  m_overlayMenu->setID("popup-overlay-menu"_spr);
+  m_overlayMenu->setID("popup-overlay-menu");
   m_overlayMenu->ignoreAnchorPointForPosition(false);
   m_overlayMenu->setPosition({ widthCS / 2.f, heightCS / 2.f });
   m_overlayMenu->setScaledContentSize(m_mainLayer->getScaledContentSize());
@@ -143,7 +151,7 @@ bool ProjectInfoPopup::setup() {
 
   m_overlayMenu->addChild(infoBtn);
 
-  log::warn("Project info popup still unfinished, please use ProjectInfoPopup::setProject to finish setting it up before displaying it");
+  AVAL_LOG_WARN("Project info popup still unfinished, please use ProjectInfoPopup::setProject to finish setting it up before displaying it");
 
   return true;
 };
@@ -155,7 +163,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
   m_avalProject = Handler::get()->GetProject(m_level->m_levelID.value());
 
   if (m_avalProject.type == Project::Type::NONE) {
-    log::error("Avalanche project type is NONE");
+    AVAL_LOG_ERROR("Avalanche project type is NONE");
     return this;
   } else if (m_avalProject.type == Project::Type::TEAM) {
     m_avalPublisher = "Avalanche";
@@ -175,7 +183,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
   fameFrame_layout->setGap(5.f);
 
   auto fameFrame = CCMenu::create();
-  fameFrame->setID("fame-frame"_spr);
+  fameFrame->setID("fame-frame");
   fameFrame->ignoreAnchorPointForPosition(false);
   fameFrame->setPosition({ m_mainLayer->getScaledContentWidth() / 2.f, m_title->getPositionY() - 20.f });
   fameFrame->setScaledContentSize({ m_mainLayer->getScaledContentWidth() * 0.75f, 12.5f });
@@ -185,7 +193,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
   m_overlayMenu->addChild(fameFrame);
 
   if (m_avalProject.fame) {
-    log::info("Project '{}' is in the Hall of Fame", m_avalProject.name);
+    AVAL_LOG_INFO("Project '{}' is in the Hall of Fame", m_avalProject.name);
 
     m_title->setFntFile("goldFont.fnt");
     m_title->setScale(1.f);
@@ -209,7 +217,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
 
     fameFrame->updateLayout(true);
   } else {
-    log::debug("Project '{}' is not in the Hall of Fame", m_avalProject.name);
+    AVAL_LOG_DEBUG("Project '{}' is not in the Hall of Fame", m_avalProject.name);
   };
 
   auto art_bottomLeft = CCSprite::createWithSpriteFrameName(m_cornerArtType.c_str());
@@ -256,15 +264,6 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
 
   m_overlayMenu->addChild(art_topRight);
 
-  auto comingSoon = CCLabelBMFont::create("More coming soon...", "bigFont.fnt");
-  comingSoon->setID("coming-soon-label");
-  comingSoon->ignoreAnchorPointForPosition(false);
-  comingSoon->setAnchorPoint({ 0.5, 0.5 });
-  comingSoon->setPosition({ m_mainLayer->getScaledContentWidth() / 2.f, m_mainLayer->getScaledContentHeight() / 2.f });
-  comingSoon->setScale(0.25f);
-
-  m_overlayMenu->addChild(comingSoon);
-
   auto hostLabelTxt = "Published by";
 
   if (m_avalProject.type == Project::Type::TEAM) hostLabelTxt = "Hosted by";
@@ -300,11 +299,35 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
     playShowcase_sprite,
     this,
     menu_selector(ProjectInfoPopup::onPlayShowcase));
-  playShowcase->setID("play-showcase-button"_spr);
+  playShowcase->setID("play-showcase-button");
   playShowcase->setPosition({ m_mainLayer->getScaledContentWidth() / 2.f, 30.f });
 
   m_overlayMenu->addChild(playShowcase_label);
   m_overlayMenu->addChild(playShowcase);
+
+  // geode settings popup button
+  auto settingsBtnSprite = CCSprite::createWithSpriteFrameName("GJ_optionsBtn_001.png");
+  settingsBtnSprite->setScale(0.75f);
+
+  auto settingsBtn = CCMenuItemSpriteExtra::create(
+    settingsBtnSprite,
+    this,
+    menu_selector(ProjectInfoPopup::settingsPopup));
+  settingsBtn->setID("settings-button");
+  settingsBtn->setPosition({ 25, 25 });
+
+  m_overlayMenu->addChild(settingsBtn);
+
+  // TODO: remove coming soon text and add more info
+
+  auto comingSoon = CCLabelBMFont::create("More coming soon...", "bigFont.fnt");
+  comingSoon->setID("coming-soon-label");
+  comingSoon->ignoreAnchorPointForPosition(false);
+  comingSoon->setAnchorPoint({ 0.5, 0.5 });
+  comingSoon->setPosition({ m_mainLayer->getScaledContentWidth() / 2.f, m_mainLayer->getScaledContentHeight() / 2.f });
+  comingSoon->setScale(0.25f);
+
+  m_overlayMenu->addChild(comingSoon);
 
   return this;
 };
@@ -315,8 +338,7 @@ void ProjectInfoPopup::show() {
   GLubyte opacity = getOpacity();
   m_mainLayer->setScale(0.1f);
 
-  m_mainLayer->runAction(
-    CCEaseElasticOut::create(CCScaleTo::create(0.3f, 1.0f), 1.6f));
+  m_mainLayer->runAction(CCEaseElasticOut::create(CCScaleTo::create(0.3f, 1.0f), 1.6f));
 
   if (!m_scene) m_scene = CCDirector::sharedDirector()->getRunningScene();
   if (!m_ZOrder) m_ZOrder = 105;